Agrega cambios de popup confirmacion eliminado

+++ src/components/Wall.js
@@ -92,7 +92,7 @@ export const Wall = () => {
             
             const btnsDelete = postComplete.querySelectorAll(".btn-delete");
             
-            const counterLikes = postComplete.querySelector(".counter-likes");
+            const counterLikes = postComplete.querySelectorAll(".counter-likes");
             const btnLike = postComplete.querySelectorAll(".btn-like");
             
             btnLike.forEach(btn => {
@@ -118,7 +118,14 @@ export const Wall = () => {
 
             btnsDelete.forEach(btn => {
                 btn.addEventListener("click", ({ target: { dataset } }) => {
-                    deleteTask(dataset.id);
+                    const result = confirm("¿Estás seguro de borrar esta publicación?");
+                    if (result === true){
+                        return deleteTask(dataset.id);
+                    }
+                    else {
+                        return false;
+                    }
+
                 }); 
             });
